apps/api/pyproject.toml
[project]
name = "convo"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "coverage>=7.11.0",
    "fastapi[standard]>=0.128.0",
    "flask>=3.1.2",
    "gunicorn>=23.0.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "PyJWT>=2.8.0",
    "pytest>=8.4.2",
    "python-multipart>=0.0.6",
    "ruff>=0.14.2",
    "uvicorn>=0.40.0",
]

[tool.ruff]
line-length = 88
target-version = "py313"
exclude = [".venv"]

# Ativa linter + formatação (substitui black/isort)
[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "W", "T20"]
ignore = ["E501", "F403"]

# Ignorar prints especificamente em arquivos de scripts
[tool.ruff.lint.per-file-ignores]
"src/infrastructure/database/init_db.py" = ["T201"]
"src/infrastructure/database/init_postgres_db.py" = ["T201"]
"src/infrastructure/database/seed_postgres_db.py" = ["T201"]
"run_*" = ["T201"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
